🐛 [fix] 로그인 버그, UI 버그 개선

+++ Cider/Cider/Presentation/ChallengeDetail/SubViews/Cell/ProgressBarCell.swift
@@ -32,8 +32,8 @@ final class ProgressBarCell: UICollectionViewCell {
 extension ProgressBarCell {
     
     func setUp(
-        averagePercent: Float,
-        myPercent: Float?
+        averagePercent: Double,
+        myPercent: Double?
     ) {
         self.backgroundColor = .white
         if let myPercent {
